+++ src/main/kotlin/com/jetbrains/snakecharm/codeInsight/SnakemakeAPI.kt
@@ -1,6 +1,10 @@
 package com.jetbrains.snakecharm.codeInsight
 
 import com.jetbrains.snakecharm.lang.SnakemakeNames
+import com.jetbrains.snakecharm.lang.SnakemakeNames.MODULE_CONFIG_KEYWORD
+import com.jetbrains.snakecharm.lang.SnakemakeNames.MODULE_META_WRAPPER_KEYWORD
+import com.jetbrains.snakecharm.lang.SnakemakeNames.MODULE_SKIP_VALIDATION_KEYWORD
+import com.jetbrains.snakecharm.lang.SnakemakeNames.MODULE_SNAKEFILE_KEYWORD
 import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_BENCHMARK
 import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_CACHE
 import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_CONDA
@@ -19,6 +23,7 @@ import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_OUTPUT
 import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_PARAMS
 import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_PRIORITY
 import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_RESOURCES
+import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_RUN
 import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_SCRIPT
 import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_SHADOW
 import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_SHELL
@@ -107,7 +112,7 @@ object SnakemakeAPI {
             SECTION_NAME,
             SECTION_HANDOVER
     )
-    val RULE_OR_CHECKPOINT_SECTION_KEYWORDS = (RULE_OR_CHECKPOINT_ARGS_SECTION_KEYWORDS + setOf(SnakemakeNames.SECTION_RUN))
+    val RULE_OR_CHECKPOINT_SECTION_KEYWORDS = (RULE_OR_CHECKPOINT_ARGS_SECTION_KEYWORDS + setOf(SECTION_RUN))
 
     /**
      * For subworkflows parsing
@@ -116,6 +121,27 @@ object SnakemakeAPI {
             SnakemakeNames.SUBWORKFLOW_WORKDIR_KEYWORD,
             SnakemakeNames.SUBWORKFLOW_SNAKEFILE_KEYWORD,
             SnakemakeNames.SUBWORKFLOW_CONFIGFILE_KEYWORD
+    )
+
+    /**
+     * For modules parsing
+     */
+    val MODULE_SECTIONS_KEYWORDS = setOf(
+        MODULE_SNAKEFILE_KEYWORD,
+        MODULE_CONFIG_KEYWORD,
+        MODULE_SKIP_VALIDATION_KEYWORD,
+        MODULE_META_WRAPPER_KEYWORD
+    )
+
+    /**
+     * For uses parsing
+     */
+    val USE_SECTIONS_KEYWORDS = RULE_OR_CHECKPOINT_SECTION_KEYWORDS - setOf(
+        SECTION_SHELL,
+        SECTION_NOTEBOOK,
+        SECTION_SCRIPT,
+        SECTION_CWL,
+        SECTION_RUN
     )
 
     /**