+++ src/main/kotlin/com/jetbrains/snakecharm/lang/parser/SmkTokenTypes.kt
@@ -13,6 +13,7 @@ object SmkTokenTypes {
     val SUBWORKFLOW_KEYWORD = PyElementType("SUBWORKFLOW_KEYWORD")
     val CHECKPOINT_KEYWORD = PyElementType("CHECKPOINT_KEYWORD") // rule
     val MODULE_KEYWORD = PyElementType("MODULE_KEYWORD")
+    val USE_KEYWORD = PyElementType("USE_KEYWORD")
 
     val WORKFLOW_CONFIGFILE_KEYWORD = PyElementType("WORKFLOW_CONFIGFILE_KEYWORD")
     val WORKFLOW_REPORT_KEYWORD = PyElementType("WORKFLOW_REPORT_KEYWORD")
@@ -58,7 +59,7 @@ object SmkTokenTypes {
     )
 
     val RULE_LIKE = TokenSet.orSet(
-            RULE_OR_CHECKPOINT, TokenSet.create(SUBWORKFLOW_KEYWORD, MODULE_KEYWORD)
+            RULE_OR_CHECKPOINT, TokenSet.create(SUBWORKFLOW_KEYWORD, MODULE_KEYWORD, USE_KEYWORD)
     )
 
     val WORKFLOW_TOPLEVEL_DECORATORS = TokenSet.orSet(