+++ src/main/kotlin/com/jetbrains/snakecharm/lang/psi/SmkPsiElements.kt
@@ -11,23 +11,22 @@ import com.jetbrains.python.psi.*
 import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.WILDCARDS_DEFINING_SECTIONS_KEYWORDS
 import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.WILDCARDS_EXPANDING_SECTIONS_KEYWORDS
 import com.jetbrains.snakecharm.codeInsight.resolve.SmkResolveUtil
-import com.jetbrains.snakecharm.lang.psi.stubs.SmkCheckpointStub
-import com.jetbrains.snakecharm.lang.psi.stubs.SmkModuleStub
-import com.jetbrains.snakecharm.lang.psi.stubs.SmkRuleStub
-import com.jetbrains.snakecharm.lang.psi.stubs.SmkSubworkflowStub
+import com.jetbrains.snakecharm.lang.psi.stubs.*
 import com.jetbrains.snakecharm.stringLanguage.lang.psi.SmkSLExpression
 
-interface SmkToplevelSection: SmkSection {
+interface SmkToplevelSection : SmkSection {
     override fun getParentRuleOrCheckPoint(): SmkRuleOrCheckpoint? = null
 }
 
-interface SmkRule: SmkRuleOrCheckpoint, StubBasedPsiElement<SmkRuleStub>
+interface SmkRule : SmkRuleOrCheckpoint, StubBasedPsiElement<SmkRuleStub>
 
-interface SmkCheckPoint: SmkRuleOrCheckpoint, StubBasedPsiElement<SmkCheckpointStub>
+interface SmkCheckPoint : SmkRuleOrCheckpoint, StubBasedPsiElement<SmkCheckpointStub>
 
-interface SmkSubworkflow: SmkRuleLike<SmkSubworkflowArgsSection>, StubBasedPsiElement<SmkSubworkflowStub>
+interface SmkSubworkflow : SmkRuleLike<SmkSubworkflowArgsSection>, StubBasedPsiElement<SmkSubworkflowStub>
 
 interface SmkModule : SmkRuleLike<SmkModuleArgsSection>, StubBasedPsiElement<SmkModuleStub>
+
+interface SmkUse : SmkRuleOrCheckpoint, StubBasedPsiElement<SmkUseStub>
 
 interface SmkRuleOrCheckpointArgsSection : SmkArgsSection, PyTypedElement { // PyNamedElementContainer
     /**
@@ -43,7 +42,7 @@ interface SmkRuleOrCheckpointArgsSection : SmkArgsSection, PyTypedElement { // P
     override fun getParentRuleOrCheckPoint(): SmkRuleOrCheckpoint = super.getParentRuleOrCheckPoint()!!
 }
 
-interface SmkSubworkflowArgsSection: SmkArgsSection {
+interface SmkSubworkflowArgsSection : SmkArgsSection {
     override fun getParentRuleOrCheckPoint(): SmkRuleOrCheckpoint? = null
 }
 
@@ -51,24 +50,24 @@ interface SmkModuleArgsSection : SmkArgsSection {
     override fun getParentRuleOrCheckPoint(): SmkRuleOrCheckpoint? = null
 }
 
-interface SmkWorkflowArgsSection: SmkArgsSection, SmkToplevelSection // PyNamedElementContainer
+interface SmkWorkflowArgsSection : SmkArgsSection, SmkToplevelSection // PyNamedElementContainer
 
-interface SmkRunSection: SmkSection, PyStatementListContainer, PyDocStringOwner {
+interface SmkRunSection : SmkSection, PyStatementListContainer, PyDocStringOwner {
     //ScopeOwner, // for control flow
 
     override fun getParentRuleOrCheckPoint(): SmkRuleOrCheckpoint = super.getParentRuleOrCheckPoint()!!
 }
 
 interface SmkWorkflowPythonBlockSection : SmkSection, SmkToplevelSection,
-        ScopeOwner, // for control flow
-        PyStatementListContainer, PyDocStringOwner
+    ScopeOwner, // for control flow
+    PyStatementListContainer, PyDocStringOwner
 
-interface SmkWorkflowLocalrulesSection: PyStatement, SmkArgsSection, SmkToplevelSection
-        // SmkArgsSection
+interface SmkWorkflowLocalrulesSection : PyStatement, SmkArgsSection, SmkToplevelSection
+// SmkArgsSection
 
-interface SmkWorkflowRuleorderSection: PyStatement, SmkArgsSection, SmkToplevelSection
+interface SmkWorkflowRuleorderSection : PyStatement, SmkArgsSection, SmkToplevelSection
 
-interface SmkReferenceExpression: PyReferenceExpression {
+interface SmkReferenceExpression : PyReferenceExpression {
     override fun getNameElement(): ASTNode? = node.findChildByType(PyTokenTypes.IDENTIFIER)
 }