+++ src/main/kotlin/com/jetbrains/snakecharm/stringLanguage/lang/SmkDSInjector.kt
@@ -1,0 +1,7 @@
+package com.jetbrains.snakecharm.stringLanguage.lang
+
+import com.jetbrains.python.documentation.doctest.PyDocstringLanguageInjector
+
+class SmkDSInjector : PyDocstringLanguageInjector() {
+
+}